feat: enhance branch catalog functionality and update user manuals
- Updated the branch catalog interface to include new filters for family and sub-family, improving item search and organization.
- Implemented sorting functionality for catalog columns, allowing users to sort items by various attributes.
- Revised JavaScript for better localization support and added internationalization features for catalog labels.
- Updated user manuals to clarify the new filtering and sorting capabilities, ensuring users understand how to navigate the catalog effectively.
- Incremented version numbers for relevant JavaScript files to reflect the latest changes.

These enhancements aim to improve user experience in the branch catalog by providing more robust search and organization tools.

+++ orders/models.py
@@ -97,8 +97,8 @@ class InternalRequestLine(models.Model):
     description = models.CharField(max_length=255)
     internal_code = models.CharField(max_length=64, blank=True)
     unit_of_measure = models.CharField(max_length=16, blank=True)
-    quantity = models.DecimalField(max_digits=12, decimal_places=3)
-    quantity_reserved = models.DecimalField(max_digits=12, decimal_places=3, default=0)
+    quantity = models.IntegerField()
+    quantity_reserved = models.IntegerField(default=0)
     unit_price = models.DecimalField(max_digits=12, decimal_places=2, default=0)
     vat_rate = models.DecimalField(max_digits=5, decimal_places=4, default=0)
     # Seam (unused in Phase 5): link a restocking PO later (lock 3).
